+++ doc/changelog.rst
@@ -1,6 +1,11 @@
 Changelog
 =========
 
+Changed
+^^^^^^^
+
+- Refactor :code:`get_field_mutability` and :code:`get_field_returability` in :code:`get_field_annotation`.
+
 [0.1.5] - 2024-06-04
 --------------------